using common.py for globals and send/receive
