+++ common.py
@@ -1,0 +1,33 @@
+import socket
+
+# Our header is 64 bytes
+PORT = 5050
+HEADER = 64
+FORMAT = 'utf-8'
+
+
+# This gets the local IPv4 address automatically without hard-coding it into the app
+SERVER = socket.gethostbyname(socket.gethostname())
+
+ADDR = (SERVER, PORT)
+
+# This is the message that we send whenever we are ready to disconnect.
+# Whenever this message is disconnected, the server will close the connection from said client.
+DISCONNECT_MESSAGE = ".exit"
+
+
+def send(msg, conn):
+    message = msg.encode(FORMAT)
+    msg_length = len(message)
+    send_length = str(msg_length).encode(FORMAT)
+    send_length += b' ' * (HEADER - len(send_length))
+    conn.send(send_length)
+    conn.send(message)
+
+
+def receive(conn):
+    msg_length = conn.recv(HEADER).decode()
+    if msg_length:
+        msg_length = int(msg_length)
+        msg = conn.recv(msg_length).decode(FORMAT)
+        return msg
